Add order service and Producer, consumer

+++ book-service/src/main/java/com/onlinebookstore/book_service/repository/InventoryRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface InventoryRepo extends JpaRepository<Inventory, Long> {
+public interface InventoryRepo extends JpaRepository<Inventory, Integer> {
     Optional<Inventory> findByBookDetails_BookId(int bookId);
 }
